+++ crates/ferroid/src/futures/mod.rs
@@ -1,9 +1,11 @@
 mod sleep_provider;
+#[cfg(feature = "snowflake")]
 mod snowflake;
 #[cfg(feature = "ulid")]
 mod ulid;
 
 pub use sleep_provider::*;
+#[cfg(feature = "snowflake")]
 pub use snowflake::*;
 #[cfg(feature = "ulid")]
 pub use ulid::*;
