+++ crates/ferroid/src/generator/mod.rs
@@ -1,9 +1,9 @@
+#[cfg(feature = "snowflake")]
 mod snowflake;
-#[cfg(test)]
-mod tests;
 #[cfg(feature = "ulid")]
 mod ulid;
 
+#[cfg(feature = "snowflake")]
 pub use snowflake::*;
 #[cfg(feature = "ulid")]
 pub use ulid::*;
