+++ crates/ferroid/src/generator/snowflake/mod.rs
@@ -2,6 +2,8 @@ mod atomic;
 mod basic;
 mod interface;
 mod lock;
+#[cfg(test)]
+mod tests;
 
 pub use atomic::*;
 pub use basic::*;
