+++ crates/ferroid/src/generator/ulid/mod.rs
@@ -1,5 +1,7 @@
 mod basic;
 mod interface;
+#[cfg(test)]
+mod tests;
 
 pub use basic::*;
 pub use interface::*;
