+++ crates/ferroid/src/id/mod.rs
@@ -1,9 +1,13 @@
 mod id;
+#[cfg(feature = "snowflake")]
 mod snowflake;
+mod to_u64;
 #[cfg(feature = "ulid")]
 mod ulid;
 
 pub use id::*;
+#[cfg(feature = "snowflake")]
 pub use snowflake::*;
+pub use to_u64::*;
 #[cfg(feature = "ulid")]
 pub use ulid::*;
