+++ crates/ferroid/src/id/snowflake.rs
@@ -1,45 +1,5 @@
-use crate::{Error, Id, Result};
+use crate::Id;
 use core::{fmt, hash::Hash};
-
-/// Trait for converting numeric-like values into a `u64`.
-///
-/// This is typically used to normalize custom duration types into milliseconds
-/// for compatibility with APIs like [`core::time::Duration::from_millis`],
-/// which are commonly required in async sleep contexts such as
-/// [`tokio::time::sleep`].
-pub trait ToU64 {
-    fn to_u64(self) -> Result<u64>;
-}
-
-impl ToU64 for u8 {
-    fn to_u64(self) -> Result<u64> {
-        Ok(self as u64)
-    }
-}
-
-impl ToU64 for u16 {
-    fn to_u64(self) -> Result<u64> {
-        Ok(self as u64)
-    }
-}
-
-impl ToU64 for u32 {
-    fn to_u64(self) -> Result<u64> {
-        Ok(self as u64)
-    }
-}
-
-impl ToU64 for u64 {
-    fn to_u64(self) -> Result<u64> {
-        Ok(self)
-    }
-}
-
-impl ToU64 for u128 {
-    fn to_u64(self) -> Result<u64> {
-        self.try_into().map_err(|_| Error::FailedToU64)
-    }
-}
 
 /// A trait representing a layout-compatible Snowflake ID generator.
 ///
