+++ crates/ferroid/src/runtime/mod.rs
@@ -3,24 +3,24 @@ mod smol;
 #[cfg(feature = "async-tokio")]
 mod tokio;
 
-#[cfg(feature = "async-smol")]
+#[cfg(all(feature = "async-smol", feature = "snowflake"))]
 mod smol_snowflake;
 #[cfg(all(feature = "async-smol", feature = "ulid"))]
 mod smol_ulid;
-#[cfg(feature = "async-tokio")]
+#[cfg(all(feature = "async-tokio", feature = "snowflake"))]
 mod tokio_snowflake;
 #[cfg(all(feature = "async-tokio", feature = "ulid"))]
 mod tokio_ulid;
 
 #[cfg(feature = "async-smol")]
 pub use smol::*;
-#[cfg(feature = "async-smol")]
+#[cfg(all(feature = "async-smol", feature = "snowflake"))]
 pub use smol_snowflake::*;
 #[cfg(all(feature = "async-smol", feature = "ulid"))]
 pub use smol_ulid::*;
 #[cfg(feature = "async-tokio")]
 pub use tokio::*;
-#[cfg(feature = "async-tokio")]
+#[cfg(all(feature = "async-tokio", feature = "snowflake"))]
 pub use tokio_snowflake::*;
 #[cfg(all(feature = "async-tokio", feature = "ulid"))]
 pub use tokio_ulid::*;
